Refactore: 유저 정보 수정 관련 mutateFunction 분리
